+++ spring-cloud-sleuth-core/src/main/java/org/springframework/cloud/sleuth/autoconfig/TraceAutoConfiguration.java
@@ -16,13 +16,10 @@
 
 package org.springframework.cloud.sleuth.autoconfig;
 
-import java.util.ArrayList;
 import java.util.Collections;
-import java.util.Comparator;
 import java.util.List;
 
 import brave.CurrentSpanCustomizer;
-import brave.ErrorParser;
 import brave.Tracer;
 import brave.Tracing;
 import brave.TracingCustomizer;
@@ -32,20 +29,8 @@ import brave.propagation.CurrentTraceContextCustomizer;
 import brave.propagation.Propagation;
 import brave.propagation.ThreadLocalCurrentTraceContext;
 import brave.sampler.Sampler;
-import io.micrometer.core.instrument.MeterRegistry;
-import org.apache.commons.logging.Log;
-import org.apache.commons.logging.LogFactory;
-import zipkin2.Span;
-import zipkin2.reporter.InMemoryReporterMetrics;
-import zipkin2.reporter.Reporter;
-import zipkin2.reporter.ReporterMetrics;
-import zipkin2.reporter.brave.ZipkinSpanHandler;
-import zipkin2.reporter.metrics.micrometer.MicrometerReporterMetrics;
 
-import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.cloud.sleuth.LocalServiceName;
@@ -85,38 +70,17 @@ public class TraceAutoConfiguration {
 	 */
 	public static final String DEFAULT_SERVICE_NAME = "default";
 
-	/**
-	 * Sort Zipkin Handlers last, so that redactions etc happen prior.
-	 */
-	static final Comparator<SpanHandler> SPAN_HANDLER_COMPARATOR = (o1, o2) -> {
-		if (o1 instanceof ZipkinSpanHandler) {
-			if (o2 instanceof ZipkinSpanHandler) {
-				return 0;
-			}
-			return 1;
-		}
-		else if (o2 instanceof ZipkinSpanHandler) {
-			return -1;
-		}
-		return 0;
-	};
-
 	@Bean
 	@ConditionalOnMissingBean
 	// NOTE: stable bean name as might be used outside sleuth
 	Tracing tracing(@LocalServiceName String serviceName, Propagation.Factory factory,
 			CurrentTraceContext currentTraceContext, Sampler sampler,
-			ErrorParser errorParser, SleuthProperties sleuthProperties,
-			@Nullable List<Reporter<zipkin2.Span>> spanReporters,
-			@Nullable List<SpanHandler> spanHandlers,
+			SleuthProperties sleuthProperties, @Nullable List<SpanHandler> spanHandlers,
 			@Nullable List<TracingCustomizer> tracingCustomizers) {
 		Tracing.Builder builder = Tracing.newBuilder().sampler(sampler)
-				.errorParser(errorParser)
 				.localServiceName(StringUtils.isEmpty(serviceName) ? DEFAULT_SERVICE_NAME
 						: serviceName)
 				.propagationFactory(factory).currentTraceContext(currentTraceContext)
-				.spanReporter(new CompositeReporter(
-						spanReporters != null ? spanReporters : Collections.emptyList()))
 				.traceId128Bit(sleuthProperties.isTraceId128())
 				.supportsJoin(sleuthProperties.isSupportsJoin());
 		if (spanHandlers != null) {
@@ -130,18 +94,7 @@ public class TraceAutoConfiguration {
 			}
 		}
 
-		reorderZipkinHandlersLast(builder);
 		return builder.build();
-	}
-
-	private void reorderZipkinHandlersLast(Tracing.Builder builder) {
-		List<SpanHandler> configuredSpanHandlers = new ArrayList<>(
-				builder.spanHandlers());
-		configuredSpanHandlers.sort(SPAN_HANDLER_COMPARATOR);
-		builder.clearSpanHandlers();
-		for (SpanHandler spanHandler : configuredSpanHandlers) {
-			builder.addSpanHandler(spanHandler);
-		}
 	}
 
 	@Bean(name = TRACER_BEAN_NAME)
@@ -184,108 +137,9 @@ public class TraceAutoConfiguration {
 
 	@Bean
 	@ConditionalOnMissingBean
-	Reporter<zipkin2.Span> noOpSpanReporter() {
-		return Reporter.NOOP;
-	}
-
-	@Bean
-	@ConditionalOnMissingBean
-	ErrorParser errorParser() {
-		return new ErrorParser();
-	}
-
-	@Bean
-	@ConditionalOnMissingBean
 	// NOTE: stable bean name as might be used outside sleuth
 	CurrentSpanCustomizer spanCustomizer(Tracing tracing) {
 		return CurrentSpanCustomizer.create(tracing);
-	}
-
-	private static final class CompositeReporter implements Reporter<zipkin2.Span> {
-
-		private static final Log log = LogFactory.getLog(CompositeReporter.class);
-
-		private final Reporter<zipkin2.Span> spanReporter;
-
-		private CompositeReporter(List<Reporter<Span>> spanReporters) {
-			this.spanReporter = spanReporters.size() == 1 ? spanReporters.get(0)
-					: new ListReporter(spanReporters);
-		}
-
-		@Override
-		public void report(Span span) {
-			this.spanReporter.report(span);
-		}
-
-		@Override
-		public String toString() {
-			return "CompositeReporter{ spanReporters=" + this.spanReporter + '}';
-		}
-
-		private static final class ListReporter implements Reporter<zipkin2.Span> {
-
-			private final List<Reporter<Span>> spanReporters;
-
-			private ListReporter(List<Reporter<Span>> spanReporters) {
-				this.spanReporters = spanReporters;
-			}
-
-			@Override
-			public void report(Span span) {
-				for (Reporter<zipkin2.Span> spanReporter : this.spanReporters) {
-					try {
-						spanReporter.report(span);
-					}
-					catch (Exception ex) {
-						log.warn("Exception occurred while trying to report the span "
-								+ span, ex);
-					}
-				}
-			}
-
-			@Override
-			public String toString() {
-				return "ListReporter{" + "spanReporters=" + this.spanReporters + '}';
-			}
-
-		}
-
-	}
-
-	@Configuration(proxyBeanMethods = false)
-	@ConditionalOnMissingClass("io.micrometer.core.instrument.MeterRegistry")
-	static class TraceMetricsInMemoryConfiguration {
-
-		@Bean
-		@ConditionalOnMissingBean
-		ReporterMetrics sleuthReporterMetrics() {
-			return new InMemoryReporterMetrics();
-		}
-
-	}
-
-	@Configuration(proxyBeanMethods = false)
-	@ConditionalOnClass(MeterRegistry.class)
-	static class TraceMetricsMicrometerConfiguration {
-
-		@Configuration(proxyBeanMethods = false)
-		@ConditionalOnMissingBean(ReporterMetrics.class)
-		static class NoReporterMetricsBeanConfiguration {
-
-			@Bean
-			@ConditionalOnBean(MeterRegistry.class)
-			ReporterMetrics sleuthMicrometerReporterMetrics(MeterRegistry meterRegistry) {
-				return MicrometerReporterMetrics.create(meterRegistry);
-			}
-
-			@Bean
-			@ConditionalOnMissingBean(MeterRegistry.class)
-			ReporterMetrics sleuthReporterMetrics() {
-				return new InMemoryReporterMetrics();
-			}
-
-		}
-
 	}
 
 }
