+++ spring-cloud-sleuth-core/src/test/java/org/springframework/cloud/sleuth/autoconfig/TraceAutoConfigurationTests.java
@@ -16,7 +16,6 @@
 
 package org.springframework.cloud.sleuth.autoconfig;
 
-import java.util.ArrayList;
 import java.util.List;
 
 import brave.Tracing;
@@ -24,9 +23,11 @@ import brave.baggage.BaggageField;
 import brave.baggage.BaggagePropagation;
 import brave.baggage.BaggagePropagationConfig.SingleBaggageField;
 import brave.baggage.BaggagePropagationCustomizer;
+import brave.handler.MutableSpan;
 import brave.handler.SpanHandler;
 import brave.propagation.B3SinglePropagation;
 import brave.propagation.Propagation;
+import brave.propagation.TraceContext;
 import brave.propagation.TraceContextOrSamplingFlags;
 import brave.sampler.RateLimitingSampler;
 import brave.sampler.Sampler;
@@ -34,75 +35,17 @@ import io.micrometer.core.instrument.MeterRegistry;
 import io.micrometer.core.instrument.simple.SimpleMeterRegistry;
 import org.assertj.core.api.BDDAssertions;
 import org.junit.jupiter.api.Test;
-import zipkin2.reporter.InMemoryReporterMetrics;
-import zipkin2.reporter.Reporter;
-import zipkin2.reporter.ReporterMetrics;
-import zipkin2.reporter.brave.ZipkinSpanHandler;
-import zipkin2.reporter.metrics.micrometer.MicrometerReporterMetrics;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.AutoConfigurations;
-import org.springframework.boot.test.context.FilteredClassLoader;
 import org.springframework.boot.test.context.runner.ApplicationContextRunner;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.Mockito.mock;
-import static org.springframework.cloud.sleuth.autoconfig.TraceAutoConfiguration.SPAN_HANDLER_COMPARATOR;
 
 public class TraceAutoConfigurationTests {
 
 	private final ApplicationContextRunner contextRunner = new ApplicationContextRunner()
 			.withConfiguration(AutoConfigurations.of(TraceAutoConfiguration.class));
-
-	@Test
-	void span_handler_comparator() {
-		SpanHandler handler1 = mock(SpanHandler.class);
-		SpanHandler handler2 = mock(SpanHandler.class);
-		ZipkinSpanHandler zipkin1 = mock(ZipkinSpanHandler.class);
-		ZipkinSpanHandler zipkin2 = mock(ZipkinSpanHandler.class);
-
-		ArrayList<SpanHandler> spanHandlers = new ArrayList<>();
-		spanHandlers.add(handler1);
-		spanHandlers.add(zipkin1);
-		spanHandlers.add(handler2);
-		spanHandlers.add(zipkin2);
-
-		spanHandlers.sort(SPAN_HANDLER_COMPARATOR);
-
-		assertThat(spanHandlers).containsExactly(handler1, handler2, zipkin1, zipkin2);
-	}
-
-	@Test
-	void should_apply_micrometer_reporter_metrics_when_meter_registry_bean_present() {
-		this.contextRunner.withUserConfiguration(WithMeterRegistry.class)
-				.run((context) -> {
-					ReporterMetrics bean = context.getBean(ReporterMetrics.class);
-
-					BDDAssertions.then(bean)
-							.isInstanceOf(MicrometerReporterMetrics.class);
-				});
-	}
-
-	@Test
-	void should_apply_in_memory_metrics_when_meter_registry_bean_missing() {
-		this.contextRunner.run((context) -> {
-			ReporterMetrics bean = context.getBean(ReporterMetrics.class);
-
-			BDDAssertions.then(bean).isInstanceOf(InMemoryReporterMetrics.class);
-		});
-	}
-
-	@Test
-	void should_apply_in_memory_metrics_when_meter_registry_class_missing() {
-		this.contextRunner.withClassLoader(new FilteredClassLoader(MeterRegistry.class))
-				.run((context) -> {
-					ReporterMetrics bean = context.getBean(ReporterMetrics.class);
-
-					BDDAssertions.then(bean).isInstanceOf(InMemoryReporterMetrics.class);
-				});
-	}
 
 	/**
 	 * Duplicates
@@ -123,8 +66,8 @@ public class TraceAutoConfigurationTests {
 	 * intentionally, to ensure configuration condition bugs do not exist.
 	 */
 	@Test
-	void should_use_RateLimitedSampler_when_reporting() {
-		this.contextRunner.withUserConfiguration(WithReporter.class).run((context -> {
+	void should_use_RateLimitedSampler_withSpanHandler() {
+		this.contextRunner.withUserConfiguration(WithSpanHandler.class).run((context -> {
 			final Sampler bean = context.getBean(Sampler.class);
 			BDDAssertions.then(bean).isInstanceOf(RateLimitingSampler.class);
 		}));
@@ -137,11 +80,10 @@ public class TraceAutoConfigurationTests {
 	 */
 	@Test
 	void should_override_sampler() {
-		this.contextRunner.withUserConfiguration(WithReporter.class, WithSampler.class)
-				.run((context -> {
-					final Sampler bean = context.getBean(Sampler.class);
-					BDDAssertions.then(bean).isSameAs(Sampler.ALWAYS_SAMPLE);
-				}));
+		this.contextRunner.withUserConfiguration(WithSampler.class).run((context -> {
+			final Sampler bean = context.getBean(Sampler.class);
+			BDDAssertions.then(bean).isSameAs(Sampler.ALWAYS_SAMPLE);
+		}));
 	}
 
 	@Test
@@ -231,21 +173,16 @@ public class TraceAutoConfigurationTests {
 	}
 
 	@Configuration
-	static class WithMeterRegistry {
+	static class WithSpanHandler {
 
 		@Bean
-		MeterRegistry meterRegistry() {
-			return new SimpleMeterRegistry();
-		}
-
-	}
-
-	@Configuration
-	static class WithReporter {
-
-		@Bean
-		Reporter<zipkin2.Span> spanReporter() {
-			return zipkin2.Span::toString;
+		SpanHandler testSpanHandler() {
+			return new SpanHandler() {
+				@Override
+				public boolean end(TraceContext context, MutableSpan span, Cause cause) {
+					return true;
+				}
+			};
 		}
 
 	}
