+++ spring-cloud-sleuth-zipkin/src/main/java/org/springframework/cloud/sleuth/zipkin2/ZipkinAutoConfiguration.java
@@ -16,24 +16,38 @@
 
 package org.springframework.cloud.sleuth.zipkin2;
 
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Comparator;
+import java.util.LinkedHashSet;
+import java.util.List;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 
+import brave.Tag;
+import brave.TracingCustomizer;
+import brave.handler.SpanHandler;
+import io.micrometer.core.instrument.MeterRegistry;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import zipkin2.CheckResult;
 import zipkin2.Span;
 import zipkin2.reporter.AsyncReporter;
+import zipkin2.reporter.InMemoryReporterMetrics;
 import zipkin2.reporter.Reporter;
 import zipkin2.reporter.ReporterMetrics;
 import zipkin2.reporter.Sender;
+import zipkin2.reporter.brave.ZipkinSpanHandler;
+import zipkin2.reporter.metrics.micrometer.MicrometerReporterMetrics;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.boot.autoconfigure.AutoConfigureBefore;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.autoconfigure.web.ServerProperties;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
@@ -45,6 +59,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 import org.springframework.core.env.Environment;
+import org.springframework.lang.Nullable;
 import org.springframework.web.client.RestTemplate;
 
 /**
@@ -73,6 +88,22 @@ import org.springframework.web.client.RestTemplate;
 public class ZipkinAutoConfiguration {
 
 	private static final Log log = LogFactory.getLog(ZipkinAutoConfiguration.class);
+
+	/**
+	 * Sort Zipkin Handlers last, so that redactions etc happen prior.
+	 */
+	static final Comparator<SpanHandler> SPAN_HANDLER_COMPARATOR = (o1, o2) -> {
+		if (o1 instanceof ZipkinSpanHandler) {
+			if (o2 instanceof ZipkinSpanHandler) {
+				return 0;
+			}
+			return 1;
+		}
+		else if (o2 instanceof ZipkinSpanHandler) {
+			return -1;
+		}
+		return 0;
+	};
 
 	/**
 	 * Zipkin reporter bean name. Name of the bean matters for supporting multiple tracing
@@ -143,6 +174,44 @@ public class ZipkinAutoConfiguration {
 		}
 	}
 
+	/** Returns one handler for as many reporters as exist. */
+	@Bean
+	SpanHandler zipkinSpanHandler(@Nullable List<Reporter<Span>> spanReporters,
+			@Nullable Tag<Throwable> errorTag) {
+		if (spanReporters == null) {
+			return SpanHandler.NOOP;
+		}
+
+		LinkedHashSet<Reporter<Span>> reporters = new LinkedHashSet<>(spanReporters);
+		reporters.remove(Reporter.NOOP);
+		if (spanReporters.isEmpty()) {
+			return SpanHandler.NOOP;
+		}
+
+		Reporter<Span> spanReporter = reporters.size() == 1 ? reporters.iterator().next()
+				: new CompositeSpanReporter(reporters.toArray(new Reporter[0]));
+
+		ZipkinSpanHandler.Builder builder = ZipkinSpanHandler.newBuilder(spanReporter);
+		if (errorTag != null) {
+			builder.errorTag(errorTag);
+		}
+		return builder.build();
+	}
+
+	/** This ensures Zipkin reporters end up after redaction, etc. */
+	@Bean
+	TracingCustomizer reorderZipkinHandlersLast() {
+		return builder -> {
+			List<SpanHandler> configuredSpanHandlers = new ArrayList<>(
+					builder.spanHandlers());
+			configuredSpanHandlers.sort(SPAN_HANDLER_COMPARATOR);
+			builder.clearSpanHandlers();
+			for (SpanHandler spanHandler : configuredSpanHandlers) {
+				builder.addSpanHandler(spanHandler);
+			}
+		};
+	}
+
 	@Bean
 	@ConditionalOnMissingBean
 	public ZipkinRestTemplateCustomizer zipkinRestTemplateCustomizer(
@@ -202,6 +271,86 @@ public class ZipkinAutoConfiguration {
 		public EndpointLocator zipkinEndpointLocator() {
 			return new DefaultEndpointLocator(this.registration, this.serverProperties,
 					this.environment, this.zipkinProperties, this.inetUtils);
+		}
+
+	}
+
+	@Configuration(proxyBeanMethods = false)
+	@ConditionalOnMissingClass("io.micrometer.core.instrument.MeterRegistry")
+	static class TraceMetricsInMemoryConfiguration {
+
+		@Bean
+		@ConditionalOnMissingBean
+		ReporterMetrics sleuthReporterMetrics() {
+			return new InMemoryReporterMetrics();
+		}
+
+	}
+
+	@Configuration(proxyBeanMethods = false)
+	@ConditionalOnClass(MeterRegistry.class)
+	static class TraceMetricsMicrometerConfiguration {
+
+		@Configuration(proxyBeanMethods = false)
+		@ConditionalOnMissingBean(ReporterMetrics.class)
+		static class NoReporterMetricsBeanConfiguration {
+
+			@Bean
+			@ConditionalOnBean(MeterRegistry.class)
+			ReporterMetrics sleuthMicrometerReporterMetrics(MeterRegistry meterRegistry) {
+				return MicrometerReporterMetrics.create(meterRegistry);
+			}
+
+			@Bean
+			@ConditionalOnMissingBean(MeterRegistry.class)
+			ReporterMetrics sleuthReporterMetrics() {
+				return new InMemoryReporterMetrics();
+			}
+
+		}
+
+	}
+
+	// Zipkin conversion only happens once per mutable span
+	static final class CompositeSpanReporter implements Reporter<Span> {
+
+		final Reporter<Span>[] reporters;
+
+		CompositeSpanReporter(Reporter<Span>[] reporters) {
+			this.reporters = reporters;
+		}
+
+		@Override
+		public void report(Span span) {
+			for (Reporter<Span> reporter : reporters) {
+				try {
+					reporter.report(span);
+				}
+				catch (RuntimeException ex) {
+					// TODO: message lifted from ListReporter: this is probably too much
+					// for warn level
+					log.warn("Exception occurred while trying to report the span " + span,
+							ex);
+				}
+			}
+		}
+
+		@Override
+		public int hashCode() {
+			return Arrays.hashCode(reporters);
+		}
+
+		@Override
+		public boolean equals(Object obj) {
+			if (!(obj instanceof CompositeSpanReporter)) {
+				return false;
+			}
+			return Arrays.equals(((CompositeSpanReporter) obj).reporters, reporters);
+		}
+
+		@Override
+		public String toString() {
+			return Arrays.toString(reporters);
 		}
 
 	}
